package for node and bundler
