+++ automerge-backend-wasm/index-bundler.js
@@ -1,0 +1,1 @@
+module.exports = require('./shared')(require('./pkg-bundler'))
