+++ automerge-backend-wasm/index-node.js
@@ -1,0 +1,1 @@
+module.exports = require('./shared')(require('./pkg-node'))
